requirements.txt
gdown
numpy==1.23.5
pandas
scikit_learn==1.2.1
scipy==1.10.0
tensorflow==2.11.0
torch==1.11.0
tqdm
vit_pytorch==0.19.6
IPython
torchvision==0.12.0
mxnet-cu112==1.9.1